backend/requirements.txt
grpcio>=1.60.0
grpcio-tools>=1.60.0
protobuf>=4.25.0
flask>=3.0.0
flask-cors>=4.0.0
psutil>=5.9.0

